Massive commit - Sensitivities required for yield curve calibration
To calibate an interest rate market composed of a number of discount
and forward curves, one needs sensitivities of assets to yield rates
so that one can form the Jacobian for the large multidimensional
root finding exercise that this becomes.

The Jacobian is the sensitivity of each target asset's present_value
to each node in each yield curve in the market being formed.
While we cannot compute these directly, we can some sensitivities
analytically.

For Vanilla Interest Rate Swaps, composed of fixed rate annuities,
and floating rate IBOR Legs, one can obtain the sensitivity of the
present value of the asset to the rates associated with the payment
and forward rate dates of the asset itself. These we call
sens_to_zero_rates

We then introduce an interpolation scheme to determine the
sensitivity of a move in a curve's node to the rates needed to price
the asset. This is done in market.rate_sensitivity.